+++ includes/object.h
@@ -6,7 +6,7 @@
 /*   By: abueskander <abueskander@student.42.fr>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/10 15:04:22 by abueskander       #+#    #+#             */
-/*   Updated: 2025/02/10 16:47:24 by abueskander      ###   ########.fr       */
+/*   Updated: 2025/02/10 17:15:49 by abueskander      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,5 +39,8 @@ float *tuplesub(float *a, float *b);
 float *tupleadd(float *a, float *b);
 float *vector(float x, float y, float z);
 float *point(float x, float y, float z);
+float *tupleSdiv(float *a, float scale);
+float *tupleSmult(float *a, float scale);
+float *tuplenegt(float *a);
 
 #endif
